Simple problem
It’s a simple problem but we need to be careful with the problem
statement.
The idea is to try all the cases from 1 to 100 flats at each floor.
Then, if it satisfies the given set of data, we compute the floor of n.
However, if there are multiple appropriate values for the floor of n,
the result will be -1.