A simple string problem
We can solve this problem in O(26*n^2) by checking all possible
substrings with a loop through an aggregating array. Because we don’t
need to take into account the order of letters in each substring, we
just need to know the appearance number of each letter in the alphabet.