A greedy problem
Because we want to minimize the number of bags used, we need to put as
many as units of garbage in each bag. Thus, on each day, we calculate
the number of bags that can be fulfilled, the residual units of garbage
will be disposed on the next day. However, there is a corner case that
we need to deal with the residual garbage from the previous day first
before moving to the next day.