1075B
For each resident, we need to know the distances from his location to
the drivers on the left and on the right. Then, we compare the two
distances and assign him to the closer driver. Thus, in this case, we
need to keep track of the position of the left and right drivers.
Initially, we find the right driver. Then, we loop through the array,
if the current resident is driver, he now becomes the left driver and
we have to update the right driver.